tf2

+++ onnx_resnet18.py
@@ -91,7 +91,7 @@ if __name__ == "__main__":
 	print("using onnxTRT fp16 mode:")
 	print("avg cost time: ", round(1000*time_sum/(i+1),4),'ms')
 	time_sum=0
-	model1=model.eval()
+	model1=model.float().eval()
 	for i in range(5):
 		inputs = torch.tensor(np.random.random((1, 3, input_size, input_size)).astype(np.float32)).cuda()
 		t1 = time.time()
@@ -117,10 +117,10 @@ if __name__ == "__main__":
 	print("avg cost time: ", round(1000.0*time_sum/5.0,4),'ms')
 	time_sum=0
 	torch.backends.cudnn.benchmark = True
-	model1=model.eval()
+	model1=model.float().eval()
 	s = torch.cuda.Stream()
 	torch.cuda.synchronize()
-	inputs = torch.tensor(np.random.random((1, 3, input_size, input_size)).astype(np.float32)).cuda()
+	inputs = torch.tensor(np.random.random((1, 3, input_size, input_size)).astype(np.float32)).cuda().float()
 	with torch.cuda.stream(s):
 		for _ in range(5):
 			out=model1(inputs)
